+++ tokyo-supermix-response-handler/src/main/java/com/tokyo/supermix/rest/response/BasicResponse.java
@@ -3,18 +3,32 @@ package com.tokyo.supermix.rest.response;
 import com.tokyo.supermix.rest.enums.RestApiResponseStatus;
 
 public class BasicResponse<T> extends ApiResponse {
+  private String message;
+  public BasicResponse(RestApiResponseStatus restApiResponseStatus) {
+    super(restApiResponseStatus);
+  }
 
-	public BasicResponse(RestApiResponseStatus restApiResponseStatus) {
-		super(restApiResponseStatus);
-	}
-	
-	public BasicResponse() {
-		super(RestApiResponseStatus.OK);
-	}
-	
-	public  BasicResponse(T responseBody, RestApiResponseStatus restApiResponseStatus, String message) {
-		super(restApiResponseStatus);
-		super.setMessage(message);
-	}
+  public BasicResponse() {
+    super(RestApiResponseStatus.OK);
+  }
+
+  public BasicResponse(RestApiResponseStatus restApiResponseStatus, String message) {
+    super(restApiResponseStatus);
+    this.setMessage(message);
+  }
+
+  public BasicResponse(T responseBody, RestApiResponseStatus restApiResponseStatus,
+      String message) {
+    super(restApiResponseStatus);
+    this.setMessage(message);
+  }
+
+  public String getMessage() {
+    return message;
+  }
+
+  public void setMessage(String message) {
+    this.message = message;
+  }
 
 }
