+++ tokyo-supermix-response-handler/src/main/java/com/tokyo/supermix/rest/response/ContentResponse.java
@@ -2,7 +2,6 @@ package com.tokyo.supermix.rest.response;
 
 import java.util.HashMap;
 import java.util.Map;
-
 import com.tokyo.supermix.rest.enums.RestApiResponseStatus;
 
 /**
@@ -10,32 +9,27 @@ import com.tokyo.supermix.rest.enums.RestApiResponseStatus;
  *
  */
 public class ContentResponse<T> extends ApiResponse {
-	public ContentResponse(RestApiResponseStatus restApiResponseStatus) {
-		super(restApiResponseStatus);
-	}
+  public ContentResponse(RestApiResponseStatus restApiResponseStatus) {
+    super(restApiResponseStatus);
+  }
 
-	public ContentResponse() {
-		super(RestApiResponseStatus.OK);
-	}
+  public ContentResponse() {
+    super(RestApiResponseStatus.OK);
+  }
 
-	private Map<String, T> results = new HashMap<String, T>();
+  private Map<String, T> results = new HashMap<String, T>();
 
-	public ContentResponse(String key, T dto, RestApiResponseStatus restApiResponseStatus) {
-		super(restApiResponseStatus);
-		results.put(key, dto);
-	}
+  public ContentResponse(String key, T dto, RestApiResponseStatus restApiResponseStatus) {
+    super(restApiResponseStatus);
+    results.put(key, dto);
+  }
 
-	public ContentResponse(T dto, RestApiResponseStatus restApiResponseStatus, String message) {
-		super(restApiResponseStatus);
-		super.setMessage(message);
-	}
+  public Map<String, T> getResults() {
+    return results;
+  }
 
-	public Map<String, T> getResults() {
-		return results;
-	}
-
-	public void setResults(Map<String, T> results) {
-		this.results = results;
-	}
+  public void setResults(Map<String, T> results) {
+    this.results = results;
+  }
 
 }
