+++ tokyo-supermix-server/src/main/java/com/tokyo/supermix/server/controller/CustomerController.java
@@ -1,0 +1,8 @@
+package com.tokyo.supermix.server.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class CustomerController {
+
+}
