+++ tokyo-supermix-server/src/main/java/com/tokyo/supermix/server/controller/SupplierCategoryController.java
@@ -24,33 +24,33 @@ import com.tokyo.supermix.util.ValidationConstance;
 @RestController
 public class SupplierCategoryController {
 
-  @Autowired
-  private SupplierCategoryService supplierCategoryService;
+	@Autowired
+	private SupplierCategoryService supplierCategoryService;
 
-  @Autowired
-  private ValidationFailureStatusCodes validationFailureStatusCodes;
+	@Autowired
+	private ValidationFailureStatusCodes validationFailureStatusCodes;
 
-  @Autowired
-  private Mapper mapper;
+	@Autowired
+	private Mapper mapper;
 
-  private static final Logger logger = Logger.getLogger(SupplierCategoryController.class);
+	private static final Logger logger = Logger.getLogger(SupplierCategoryController.class);
 
-  @PostMapping(value = EndpointURI.SUPPLIER_CATEGORY)
-  public ResponseEntity<Object> createSupplierCategory(
-      @RequestBody SupplierCategoryDto supplierCategoryDto) {
-    if (supplierCategoryService.isSupplierCategoryAlreadyExist(supplierCategoryDto.getCategory())) {
-      logger.debug("Supplier Category already exists: createSupplierCategory(), category: {}");
-      return new ResponseEntity<>(new BasicResponse<>(
-          new ValidationFailure(Constants.SUPPLIER_CATEGORY_NAME,
-              validationFailureStatusCodes.getSupplierCategoryAlreadyExist()),
-          RestApiResponseStatus.VALIDATION_FAILURE, ValidationConstance.SUPPLIER_CATEGORY_EXIST),
-          HttpStatus.BAD_REQUEST);
-    }
-    SupplierCategory supplierCategory = mapper.map(supplierCategoryDto, SupplierCategory.class);
-    supplierCategoryService.createSupplierCategory(supplierCategory);
-    return new ResponseEntity<>(
-        new ApiResponse(RestApiResponseStatus.CREATED, Constants.ADD_SUPPLIER_CATEGORY_SUCCESS),
-        HttpStatus.OK);
+	@PostMapping(value = EndpointURI.SUPPLIER_CATEGORY)
+	public ResponseEntity<Object> createSupplierCategory(@RequestBody SupplierCategoryDto supplierCategoryDto) {
+		if (supplierCategoryService.isSupplierCategoryAlreadyExist(supplierCategoryDto.getCategory())) {
+			logger.debug("Supplier Category already exists: createSupplierCategory(), category: {}");
+			return new ResponseEntity<>(
+					new BasicResponse<>(
+							new ValidationFailure(Constants.SUPPLIER_CATEGORY_NAME,
+									validationFailureStatusCodes.getSupplierCategoryAlreadyExist()),
+							RestApiResponseStatus.VALIDATION_FAILURE, ValidationConstance.SUPPLIER_CATEGORY_EXIST),
+					HttpStatus.BAD_REQUEST);
+		}
+		SupplierCategory supplierCategory = mapper.map(supplierCategoryDto, SupplierCategory.class);
+		supplierCategoryService.createSupplierCategory(supplierCategory);
+		return new ResponseEntity<>(
+				new BasicResponse<>(RestApiResponseStatus.CREATED, Constants.ADD_SUPPLIER_CATEGORY_SUCCESS),
+				HttpStatus.OK);
 
-  }
+	}
 }
